requirements.txt
faster-whisper
torch
pyannote.audio
pyannote.core
numpy
websockets
webrtcvad
